# apps/syn-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Syntropic137 API",
    "description": "API for Syntropic137. Provides real-time observability for workflow execution, agent sessions, and artifacts.",
    "version": "0.5.1"
  },
  "paths": {
    "/workflows": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows Endpoint",
        "description": "List all workflow templates.",
        "operationId": "list_workflows_endpoint_workflows_get",
        "parameters": [
          {
            "name": "workflow_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by workflow type",
              "title": "Workflow Type"
            },
            "description": "Filter by workflow type"
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include archived workflows",
              "default": false,
              "title": "Include Archived"
            },
            "description": "Include archived workflows"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort field (- prefix = descending)",
              "title": "Order By"
            },
            "description": "Sort field (- prefix = descending)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow Endpoint",
        "description": "Create a new workflow template.",
        "operationId": "create_workflow_endpoint_workflows_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow Endpoint",
        "description": "Get workflow details by ID (supports partial ID prefix matching).",
        "operationId": "get_workflow_endpoint_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workflows"
        ],
        "summary": "Archive (soft-delete) a workflow template",
        "description": "Archive (soft-delete) a workflow template.\n\nArchived templates are excluded from listing by default but remain\naccessible via `GET /workflows/{id}` and with `?include_archived=true`.",
        "operationId": "delete_workflow_endpoint_workflows__workflow_id__delete",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkflowResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workflow template not found"
          },
          "409": {
            "description": "Conflict \u2014 workflow has active executions or is already archived"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/workflows/validate": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Validate Yaml Endpoint",
        "description": "Validate a workflow YAML definition.",
        "operationId": "validate_yaml_endpoint_workflows_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateYamlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateYamlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}/phases/{phase_id}": {
      "put": {
        "tags": [
          "workflows"
        ],
        "summary": "Update Phase Prompt Endpoint",
        "description": "Update a workflow phase's prompt template and optional config.",
        "operationId": "update_phase_prompt_endpoint_workflows__workflow_id__phases__phase_id__put",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "phase_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Phase Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhasePromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePhaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkflowRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1,
                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type",
            "default": "custom"
          },
          "classification": {
            "type": "string",
            "title": "Classification",
            "default": "standard"
          },
          "repository_url": {
            "type": "string",
            "title": "Repository Url",
            "default": "https://github.com/example/repo"
          },
          "repository_ref": {
            "type": "string",
            "title": "Repository Ref",
            "default": "main"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "phases": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phases"
          },
          "input_declarations": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Declarations"
          },
          "repos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Repos",
            "description": "Default GitHub URLs for this workflow template (ADR-058). Can be overridden at execution time via the repos field on the execute request."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateWorkflowRequest"
      },
      "CreateWorkflowResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "workflow_type",
          "status"
        ],
        "title": "CreateWorkflowResponse"
      },
      "DeleteWorkflowResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "status"
        ],
        "title": "DeleteWorkflowResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdatePhaseResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "phase_id": {
            "type": "string",
            "title": "Phase Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "phase_id",
          "status"
        ],
        "title": "UpdatePhaseResponse"
      },
...
      "WorkflowListResponse": {
        "properties": {
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSummaryResponse"
            },
            "type": "array",
            "title": "Workflows"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "workflows",
          "total"
        ],
        "title": "WorkflowListResponse"
      },
      "WorkflowResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "classification": {
            "type": "string",
            "title": "Classification"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/PhaseDefinition"
            },
            "type": "array",
            "title": "Phases"
          },
          "input_declarations": {
            "items": {
              "$ref": "#/components/schemas/InputDeclarationModel"
            },
            "type": "array",
            "title": "Input Declarations"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "runs_count": {
            "type": "integer",
            "title": "Runs Count",
            "default": 0
          },
          "runs_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runs Link"
          },
          "repository_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Url"
          },
          "repos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Repos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "workflow_type",
          "classification"
        ],
        "title": "WorkflowResponse"
      },
...
    }
  },
...
}